Implement resolving members
This implements resolving members, but doesn't handle scoping for them.
Handling scoping for members requires further investigation on how do we actually want to handle it to achieve the largest ammount of common code between languages.
Scoping for members would require deep resolving indentifiers, to find the underlying type, but this process is very different when dealing with statically typed language such as C#, dynamically typed langauge such as lua, or structurally typed language such as TS.
In short - implementing scoping for members is postponed until we implement more languages.